chore: configured add menu
