+++ backend/src/controller/shop.controller.js
@@ -1,31 +1,11 @@
 import { Shop } from "../models/shop.model.js";
-import { minioClient, bucketName } from "../config/minio.config.js";
-import fs from "fs";
-import util from "util";
-
-const unlinkFile = util.promisify(fs.unlink);
+import { uploadFile } from "../service/minio.service.js";
 
 export const createShop = async (req, res) => {
   try {
     const { shopName, address, contactNo, ownerId, email } = req.body;
 
-    if (!req.file) {
-      return res.status(400).json({
-        message: "No file uploaded",
-      });
-    }
-
-    const file = req.file;
-    const objectName = file.filename;
-
-    const bucketExists = await minioClient.bucketExists(bucketName);
-    if (!bucketExists) {
-      console.log(`Bucket '${bucketName}' does not exist.`);
-      return res.status(404).send(`Bucket '${bucketName}' does not exist.`);
-    }
-
-    await minioClient.fPutObject(bucketName, objectName, file.path);
-    const pictureUrl = `${process.env.MINIO_ENDPOINT}/${bucketName}/${objectName}`;
+    const pictureUrl = await uploadFile(req.file);
 
     const newShop = new Shop({
       shopName,
@@ -37,8 +17,6 @@ export const createShop = async (req, res) => {
     });
 
     const savedShop = await newShop.save();
-
-    await unlinkFile(file.path);
 
     res.status(201).json({
       message: "Shop created successfully!",
