+++ backend/src/routes/menu.route.js
@@ -6,10 +6,11 @@ import {
   deleteItem,
   getAllMenuOfShop,
 } from "../controller/menu.controller.js";
+import { upload } from "../middleware/multer.middleware.js";
 
 const router = Router();
 
-router.route("/:shopId/menu").post(createMenu);
+router.route("/:shopId/menu").post(upload.single("picture"), createMenu);
 router.route("/:shopId/menu").get(getAllMenuOfShop);
 router.route("/menu/:menuId/item/:itemId").get(getMenuItem);
 router.route("/:shopId/menu/:menuId/item/:itemId").patch(updateMenu);
